refactor: Restructure the routing system
