+++ src/modules/shared/application/config.ts
@@ -1,0 +1,5 @@
+const REST_API_PORT = process.env.REST_API_PORT || 3000
+
+export const config = Object.freeze({
+  app: { REST_API_PORT },
+})
