+++ src/routes/ROUTES_ENUM.ts
@@ -1,0 +1,7 @@
+export const ROUTES_ENUM = Object.freeze({
+  v1: {
+    DOCS: 'api/v1/api-docs',
+    // Add more routes here if you need to reference them in multiple places
+    // or if another route depends on ir
+  },
+})
